+++ rozofs/rpc/sclient.c
@@ -57,7 +57,7 @@ void sclient_release(sclient_t * clt) {
         rpcclt_release(&clt->rpcclt);
 }
 
-int sclient_write(sclient_t * clt, cid_t cid, sid_t sid, uint8_t layout,
+int sclient_write(sclient_t * clt, cid_t cid, sid_t sid, uint32_t layout,
         uint8_t spare, sid_t dist_set[ROZOFS_SAFE_MAX], fid_t fid, bid_t bid,
         uint32_t nb_proj, const bin_t * bins) {
     int status = -1;
@@ -100,7 +100,7 @@ out:
     return status;
 }
 
-int sclient_read(sclient_t * clt, cid_t cid, sid_t sid, uint8_t layout, uint8_t spare,
+int sclient_read(sclient_t * clt, cid_t cid, sid_t sid, uint32_t layout, uint8_t spare,
         sid_t dist_set[ROZOFS_SAFE_MAX], fid_t fid, bid_t bid,
         uint32_t nb_proj, bin_t * bins) {
     int status = -1;
@@ -145,7 +145,7 @@ out:
     return status;
 }
 
-int sclient_read_rbs(sclient_t * clt, cid_t cid, sid_t sid, uint8_t layout,
+int sclient_read_rbs(sclient_t * clt, cid_t cid, sid_t sid, uint32_t layout,
         uint8_t spare, sid_t dist_set[ROZOFS_SAFE_MAX], fid_t fid, bid_t bid,
         uint32_t nb_proj, uint32_t * nb_proj_recv, bin_t * bins) {
     int status = -1;
@@ -205,7 +205,7 @@ out:
 
 // XXX Never used yet
 
-int storageclt_truncate(sclient_t * clt, cid_t cid, sid_t sid, uint8_t layout,
+int storageclt_truncate(sclient_t * clt, cid_t cid, sid_t sid, uint32_t layout,
         uint8_t spare, sid_t dist_set[ROZOFS_SAFE_MAX], fid_t fid,
         tid_t proj_id, bid_t bid) {
     int status = -1;
