+++ rozofs/rpc/sproto.h
@@ -34,7 +34,7 @@ typedef struct sp_status_ret_t sp_status_ret_t;
 struct sp_write_arg_t {
 	uint16_t cid;
 	uint8_t sid;
-	uint8_t layout;
+	uint32_t layout;
 	uint8_t spare;
 	uint32_t dist_set[ROZOFS_SAFE_MAX_RPC];
 	sp_uuid_t fid;
@@ -52,7 +52,7 @@ typedef struct sp_write_arg_t sp_write_arg_t;
 struct sp_write_arg_no_bins_t {
 	uint16_t cid;
 	uint8_t sid;
-	uint8_t layout;
+	uint32_t layout;
 	uint8_t spare;
 	uint32_t dist_set[ROZOFS_SAFE_MAX_RPC];
 	sp_uuid_t fid;
@@ -67,7 +67,7 @@ typedef struct sp_write_arg_no_bins_t sp_write_arg_no_bins_t;
 struct sp_read_arg_t {
 	uint16_t cid;
 	uint8_t sid;
-	uint8_t layout;
+	uint32_t layout;
 	uint8_t spare;
 	uint32_t dist_set[ROZOFS_SAFE_MAX_RPC];
 	sp_uuid_t fid;
@@ -79,7 +79,7 @@ typedef struct sp_read_arg_t sp_read_arg_t;
 struct sp_truncate_arg_t {
 	uint16_t cid;
 	uint8_t sid;
-	uint8_t layout;
+	uint32_t layout;
 	uint8_t spare;
 	uint8_t dist_set[ROZOFS_SAFE_MAX];
 	sp_uuid_t fid;
