+++ rozofs/rpc/storcli_proto.h
@@ -35,7 +35,7 @@ struct storcli_write_arg_t {
 	uint16_t cid;
 	uint8_t sid;
 	uint8_t empty_file;
-	uint8_t layout;
+	uint32_t layout;
 	uint8_t dist_set[ROZOFS_SAFE_MAX];
 	storcli_uuid_t fid;
 	uint64_t off;
@@ -50,7 +50,7 @@ struct storcli_write_arg_no_data_t {
 	uint16_t cid;
 	uint8_t sid;
 	uint8_t empty_file;
-	uint8_t layout;
+	uint32_t layout;
 	uint8_t dist_set[ROZOFS_SAFE_MAX];
 	storcli_uuid_t fid;
 	uint64_t off;
@@ -61,7 +61,7 @@ typedef struct storcli_write_arg_no_data_t storcli_write_arg_no_data_t;
 struct storcli_read_arg_t {
 	uint16_t cid;
 	uint8_t sid;
-	uint8_t layout;
+	uint32_t layout;
 	uint8_t spare;
 	uint8_t dist_set[ROZOFS_SAFE_MAX];
 	storcli_uuid_t fid;
@@ -74,7 +74,7 @@ typedef struct storcli_read_arg_t storcli_read_arg_t;
 struct storcli_truncate_arg_t {
 	uint16_t cid;
 	uint8_t sid;
-	uint8_t layout;
+	uint32_t layout;
 	uint8_t spare;
 	uint8_t dist_set[ROZOFS_SAFE_MAX];
 	storcli_uuid_t fid;
