+++ src/storaged/rbs_sclient.c
@@ -197,7 +197,7 @@ out:
 }
 
 int rbs_read_proj(sclient_t *storage, cid_t cid, sid_t sid, uint8_t stor_idx,
-        uint8_t layout, sid_t dist_set[ROZOFS_SAFE_MAX], fid_t fid,
+        uint32_t layout, sid_t dist_set[ROZOFS_SAFE_MAX], fid_t fid,
         bid_t first_block_idx, uint32_t nb_blocks_2_read,
         uint32_t * nb_blocks_read, rbs_projection_ctx_t * proj_ctx_p) {
 
@@ -255,7 +255,7 @@ typedef struct rbs_blocks_recv_ctx {
 
 rbs_blocks_recv_ctx_t rbs_blocks_recv_tb[ROZOFS_SAFE_MAX];
 
-static int rbs_read_proj_set(sclient_t **storages, uint8_t layout, cid_t cid,
+static int rbs_read_proj_set(sclient_t **storages, uint32_t layout, cid_t cid,
         sid_t dist_set[ROZOFS_SAFE_MAX], fid_t fid, bid_t first_block_idx,
         uint32_t nb_blocks_2_read, uint32_t * nb_blocks_read, int retry_nb,
         rbs_storcli_ctx_t * working_ctx_p) {
@@ -356,7 +356,7 @@ out:
     return status;
 }
 
-int rbs_read_blocks(sclient_t **storages, uint8_t layout, cid_t cid,
+int rbs_read_blocks(sclient_t **storages, uint32_t layout, cid_t cid,
         sid_t dist_set[ROZOFS_SAFE_MAX], fid_t fid, bid_t first_block_idx,
         uint32_t nb_blocks_2_read, uint32_t * nb_blocks_read, int retry_nb,
         rbs_storcli_ctx_t * working_ctx_p) {
