+++ src/storaged/rbs_transform.c
@@ -34,7 +34,7 @@ angle_t rbs_angles[ROZOFS_SAFE_MAX];
 uint16_t rbs_psizes[ROZOFS_SAFE_MAX];
 uint8_t rbs_prj_idx_table[ROZOFS_SAFE_MAX];
 
-int rbs_check_timestamp_tb(rbs_projection_ctx_t *prj_ctx_p, uint8_t layout,
+int rbs_check_timestamp_tb(rbs_projection_ctx_t *prj_ctx_p, uint32_t layout,
         uint32_t block_idx, uint8_t *prj_idx_tb_p, uint64_t *timestamp_p,
         uint16_t *effective_length_p) {
 
@@ -145,7 +145,7 @@ int rbs_check_timestamp_tb(rbs_projection_ctx_t *prj_ctx_p, uint8_t layout,
     return -1;
 }
 
-int rbs_transform_inverse(rbs_projection_ctx_t *prj_ctx_p, uint8_t layout,
+int rbs_transform_inverse(rbs_projection_ctx_t *prj_ctx_p, uint32_t layout,
         uint32_t first_block_idx, uint32_t number_of_blocks,
         rbs_inverse_block_t *block_ctx_p, char *data) {
 
@@ -246,7 +246,7 @@ int rbs_transform_inverse(rbs_projection_ctx_t *prj_ctx_p, uint8_t layout,
 }
 
 int rbs_transform_forward_one_proj(rbs_projection_ctx_t * prj_ctx_p,
-        rbs_inverse_block_t * block_ctx_p, uint8_t layout,
+        rbs_inverse_block_t * block_ctx_p, uint32_t layout,
         uint32_t first_block_idx, uint32_t number_of_blocks,
         tid_t projection_id, char *data) {
